pyproject.toml
[project]
name = "babycam-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.2.0",
    "channels-redis>=4.2.1",
    "daphne>=4.1.2",
    "django>=5.1.6",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "numpy>=2.2.2",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "whitenoise>=6.8.2",
]
